test(factory): keep artifact projection coverage internal
